+++ src/components/templates/BaseSection/сomposables/entity.ts
@@ -1,0 +1,12 @@
+import { convertCamelCase } from '../../../../helpers'
+
+export const generateEntityUrl = (entityName: string): string => {
+  const indexSymbolNextDash = entityName.indexOf('-') + 1
+
+  const entityNameForLoad = entityName.replace(
+    entityName[indexSymbolNextDash],
+    entityName[indexSymbolNextDash].toLowerCase(),
+  )
+
+  return convertCamelCase(entityNameForLoad, '/')
+}
